Sprt test: Axiom-v1.0.2 vs Axiom-v1.0.1 Alpha-beta pruning
NOTE: I tried this originally with testing for a beta cutoff befor alpha was raised. This caused a elo loss of about -35 +/- 20. Then I tried with MVV-LVA too and it turned to -120 +/- 50. The issue was that it was not updating the best move in the root node when a beta cutoff occurred.

In addition, I added a qSearch function (unused) and a capture generator (unused)

Score of Axiom_new vs Axiom_old: 82 - 13 - 27  [0.783] 122
...      Axiom_new playing White: 48 - 6 - 7  [0.844] 61
...      Axiom_new playing Black: 34 - 7 - 20  [0.721] 61
...      White vs Black: 55 - 40 - 27  [0.561] 122
Elo difference: 222.7 +/- 62.6, LOS: 100.0 %, DrawRatio: 22.1 %
SPRT: llr 2.92 (101.0%), lbound -2.25, ubound 2.89 - H1 was accepted

Player: Axiom_new
   "Draw by 3-fold repetition": 3
   "Draw by fifty moves rule": 22
   "Draw by insufficient mating material": 2
   "Loss: Black mates": 6
   "Loss: White mates": 7
   "No result": 5
   "Win: Black mates": 34
   "Win: White mates": 48
Player: Axiom_old
   "Draw by 3-fold repetition": 3
   "Draw by fifty moves rule": 22
   "Draw by insufficient mating material": 2
   "Loss: Black mates": 34
   "Loss: White mates": 48
   "No result": 5
   "Win: Black mates": 6
   "Win: White mates": 7
Finished match